fix break loop
